+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Added syntax highlighting for gamedata files.
  - Added syntax highlighting for cfg files.
  - Added a fileicon for .cfg files.
+ - Added linter and builtin compiler options.
 
 ### 1.10.2
  - Added link to documentation in hover help.